transfer Extras now finished

+++ src/main/java/org/example/yalla_api/admin/dtos/transfer/TransferExtraDTO.java
@@ -9,7 +9,7 @@ public class TransferExtraDTO {
 
     private Long id;
 
-    private Integer paxValue; // Pax equivalent value
+    private Double paxValue; // Pax equivalent value
 
     private List<TransferExtraTranslationDTO> translations;
